PDFBOX-4883: replace doubleValue with floatValue
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1878876 13f79535-47bb-0310-9956-ffa450edef68

+++ preflight/src/main/java/org/apache/pdfbox/preflight/parser/PreflightParser.java
@@ -611,7 +611,7 @@ public class PreflightParser extends PDFParser
             COSNumber number = (COSNumber) result;
             if (number instanceof COSFloat)
             {
-                Double real = number.doubleValue();
+                float real = number.floatValue();
                 if (real > MAX_POSITIVE_FLOAT || real < MAX_NEGATIVE_FLOAT)
                 {
                     addValidationError(new ValidationError(ERROR_SYNTAX_NUMERIC_RANGE,
